server refactor part 1
redesigned server messaging. only one call can be made to server / to
client. this call contains a payload and a message type.

Ideas for tomorrow: add more to the WebClient Translator. The
PlayerConnection class will end up being really minimal. Simply acting
as a glue between the web and the server.